Manual

Introduction

General interface

TransformVariables.dimension — Function.

dimension(t::TransformReals)

The dimension (number of elements) that t transforms.

source

TransformVariables.transform — Function.

transform(t, x)

Transform x using t.

source

TransformVariables.transform_and_logjac — Function.

transform_and_logjac(t, x)

Transform x using t; calculating the log Jacobian determinant, returned as the second value.

source

TransformVariables.inverse — Function.

inverse(t::TransformReals, y)

Return x so that transform(t, x) ≈ y.

source

TransformVariables.transform_logdensity — Function.

transform_logdensity(t, f, x)

Let $y = t(x)$, and $f(y)$ a log density at y. This function evaluates f ∘ t as a log density, taking care of the log Jacobian correction.

source

Specific transformations

Scalar transforms

TransformVariables.to_interval — Function.

to_interval(left, right)

Return a transformation that transforms a single real number to the given (open) interval.

left < right is required, but may be -∞ or ∞, respectively, in which case the appropriate transformation is selected. See ∞.

Some common transformations are predefined as constants, see to_ℝ, to_ℝ₋, to_ℝ₊, to_𝕀.

source

TransformVariables.∞ — Constant.

Placeholder representing of infinity for specifing interval boundaries. Supports the - operator, ie -∞.

source

TransformVariables.to_ℝ — Constant.

Transform to the real line (identity).

source

TransformVariables.to_ℝ₊ — Constant.

Transform to a non-negative real number.

source

TransformVariables.to_ℝ₋ — Constant.

Transform to a non-positive real number.

source

TransformVariables.to_𝕀 — Constant.

Transform to the unit interval (0, 1).

source

Special arrays

TransformVariables.to_unitvec — Function.

to_unitvec(n)

Return a transformation that transforms n - 1 real numbers to a unit vector (under Euclidean norm).

source

TransformVariables.to_corr_cholesky — Function.

to_corr_cholesky(n)

Return a transformation that transforms real numbers to an $n×n$ upper-triangular matrix Ω, such that Ω'*Ω is a correlation matrix (positive definite, with unit diagonal).

source

Aggregation of transformations

TransformVariables.to_array — Function.

to_array(transformation, dims)

Return a transformation that applies transformation repeatedly to create an array with the given dims.

source

TransformVariables.to_tuple — Function.

to_tuple(transformations)

Return a transformation that transforms consecutive groups of real numbers to a (named) tuple, using the given transformations.

source

to_tuple(transformations)

source

to_tuple(transformations)

source

to_tuple(; (:: transformations NTransformations)...)

source

Defining custom transformations

TransformVariables.logjac_forwarddiff — Function.

logjac_forwarddiff(transformation, flatten, x)

Calculate the log Jacobian of transformation at x using ForwardDiff.jacobian.

flatten should be a bijection that maps the result of the transformation to a vector of reals. This means that elements which are redundant should not be part of the result; since f is continuous, this means that flatten should have the same number of elements as x.

source

TransformVariables.CustomTransform — Type.

CustomTransform(dimension, transformation, flatten)

Wrap a custom transform y = transformation(x) in a type that calculates the log Jacobian using ForwardDiff when necessary. See [logjac_forwarddiff] for the documentation of flatten.

source