Internals

Types for various transformations

These are not part of the API, use the as constructor or one of the predefined constants.

Scalar transformations

TransformVariables.Identity — Type

struct Identity <: TransformVariables.ScalarTransform

Identity $x ↦ x$.

TransformVariables.ScaledShiftedLogistic — Type

struct ScaledShiftedLogistic{T<:Real} <: TransformVariables.ScalarTransform

Maps to (scale, shift + scale) using x ↦ logistic(x)*scale + shift.

TransformVariables.ShiftedExp — Type

struct ShiftedExp{D, T<:Real} <: TransformVariables.ScalarTransform

Shifted exponential. When D::Bool == true, maps to (shift, ∞) using x ↦ shift + eˣ, otherwise to (-∞, shift) using x ↦ shift - eˣ.

Aggregating transformations

TransformVariables.ArrayTransform — Type

struct ArrayTransform{T<:TransformVariables.AbstractTransform, M} <: TransformVariables.VectorTransform

Apply transformation repeatedly to create an array with given dims.

TransformVariables.TransformTuple — Type

struct TransformTuple{T} <: TransformVariables.VectorTransform

Transform consecutive groups of real numbers to a tuple, using the given transformations.

Wrapper for inverse

Missing docstring.

Missing docstring for TransformVariables.InverseTransform. Check Documenter's build log for details.

Types and type aliases

TransformVariables.AbstractTransform — Type

abstract type AbstractTransform

Supertype for all transformations in this package.

Interface

The user interface consists of

dimension

transform

transform_and_logjac

[inverse]@(ref), inverse!

inverse_eltype.

TransformVariables.ScalarTransform — Type

abstract type ScalarTransform <: TransformVariables.AbstractTransform

Transform a scalar (real number) to another scalar.

Subtypes mustdefine transform, transform_and_logjac, and inverse; other methods of of the interface should have the right defaults.

TransformVariables.VectorTransform — Type

abstract type VectorTransform <: TransformVariables.AbstractTransform

Transformation that transforms <: AbstractVectors to other values.

Implementation

Implements transform and transform_and_logjac via transform_with, and inverse via inverse!.

Conditional calculation of log Jacobian determinant

TransformVariables.LogJacFlag — Type

abstract type LogJacFlag

Flag used internally by the implementation of transformations, as explained below.

When calculating the log jacobian determinant for a matrix, initialize with

logjac_zero(flag, x)

and then accumulate with log jacobians as needed with +.

When flag is LogJac, methods should return the log Jacobian as the second argument, otherwise NoLogJac, which simply combines to itself with +, serving as an empty placeholder. This allows methods to share code of the two implementations.

TransformVariables.LogJac — Type

Calculate log Jacobian as the second value.

TransformVariables.NoLogJac — Type

Don't calculate log Jacobian, return NOLOGJAC as the second value.

TransformVariables.logjac_zero — Function

logjac_zero(_, T)

Initial value for log Jacobian calculations.

Helper functions

TransformVariables.transform_with — Function

transform_with(flag::LogJacFlag, transformation, x::AbstractVector, index)

Transform elements of x from index, using transformation.

Return (y, logjac), index′, where

y is the result of the transformation,

logjac is the the log Jacobian determinant or a placeholder, depending on flag,

index′ is the next index in x after the elements used for the transformation

Internal function. Implementations

can assume that x has enough elements for transformation (ie @inbounds can be

used),

should work with generalized indexing on x.

TransformVariables._transform_tuple — Function

_transform_tuple(flag, x, index, _)

Helper function for transforming tuples. Used internally, to help type inference. Use via transfom_tuple.

TransformVariables._inverse!_tuple — Function

_inverse!_tuple(x, index, ts, ys)

Helper function for inverting tuples of transformations. Used internally.

Performs no argument validation, caller should do this.

TransformVariables._inverse_eltype_tuple — Function

_inverse_eltype_tuple(ts, ys)

Helper function determining element type of inverses from tuples. Used internally.

Performs no argument validation, caller should do this.

TransformVariables.unit_triangular_dimension — Function

unit_triangular_dimension(n)

Number of elements (strictly) above the diagonal in an $n×n$ matrix.

Building blocks for transformations

TransformVariables.l2_remainder_transform — Function

(y, r, ℓ) =

l2_remainder_transform(flag, x, r)

Given $x ∈ ℝ$ and $0 ≤ r ≤ 1$, return (y, r′) such that

$y² + r′² = r²$,

$y: |y| ≤ r$ is mapped with a bijection from x.

ℓ is the log Jacobian (whether it is evaluated depends on flag).

TransformVariables.l2_remainder_inverse — Function

(x, r′) =

l2_remainder_inverse(y, r)

Inverse of l2_remainder_transform in x and y.