Manual

Introduction

General interface

TransformVariables.dimension — Function.

dimension(t::AbstractTransform)

The dimension (number of elements) that t transforms.

Types should implement this method.

source

TransformVariables.transform — Function.

transform(t, x)

Transform x using t.

source

TransformVariables.transform_and_logjac — Function.

transform_and_logjac(t, x)

Transform x using t; calculating the log Jacobian determinant, returned as the second value.

source

TransformVariables.inverse — Function.

inverse(t::AbstractTransform, y)

Return x so that transform(t, x) ≈ y.

inverse(t)

Return a callable equivalen to y -> inverse(t, y).

source

TransformVariables.transform_logdensity — Function.

transform_logdensity(t, f, x)

Let $y = t(x)$, and $f(y)$ a log density at y. This function evaluates f ∘ t as a log density, taking care of the log Jacobian correction.

source

TransformVariables.as — Function.

as(T, args...)

Shorthand for constructing transformations with image in T. args determines or modifies behavior, details depend on T.

Not all transformations have an as method, some just have direct constructors. See methods(as) for a list.

Examples

as(Real, -∞, 1)     # transform a real number to (-∞, 1)
as(Array, 10, 2)    # reshape 20 real numbers to a 10x2 matrix
as((a = ℝ₊, b = 𝕀)) # transform 2 real numbers a NamedTuple, with a > 0, 0 < b < 1

source

Specific transformations

Scalar transforms

TransformVariables.∞ — Constant.

Placeholder representing of infinity for specifing interval boundaries. Supports the - operator, ie -∞.

source

TransformVariables.ℝ — Constant.

Transform to the real line (identity).

source

TransformVariables.ℝ₊ — Constant.

Transform to a non-negative real number.

source

TransformVariables.ℝ₋ — Constant.

Transform to a non-positive real number.

source

TransformVariables.𝕀 — Constant.

Transform to the unit interval (0, 1).

source

Special arrays

TransformVariables.UnitVector — Type.

UnitVector(n)

Transform n-1 real numbers to a unit vector of length n, under the Euclidean norm.

source

TransformVariables.CorrCholeskyFactor — Type.

CorrCholeskyFactor(n)

Cholesky factor of a correlation matrix of size n.

source

Aggregation of transformations

FIXME explain as syntax

Defining custom transformations

TransformVariables.logjac_forwarddiff — Function.

logjac_forwarddiff(f, x)

Calculate the log Jacobian determinant of f at x using `ForwardDiff.

Note

f should be a bijection, mapping from vectors of real numbers to vectors of equal length.

source

TransformVariables.value_and_logjac_forwarddiff — Function.

value_and_logjac_forwarddiff(f, x)
value_and_logjac_forwarddiff(f, x, flatten)

Calculate the value and the log Jacobian determinant of f at x. flatten is used to get a vector out of the result that makes f a bijection.

source

TransformVariables.CustomTransform — Type.

CustomTransform(g, f, flatten)

Wrap a custom transform y = f(transform(g, x))in a type that calculates the log Jacobian of∂y/∂xusingForwardDiff` when necessary.

Usually, g::TransformReals, but when an integer is used, it amounts to the identity transformation with that dimension.

flatten should take the result from f, and return a flat vector with no redundant elements, so that $x ↦ y$ is a bijection. For example, for a covariance matrix the elements below the diagonal should be removed.

source